Make user password is hashed using bcrypt
