+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@hapi/hapi": "^20.2.1",
+    "bcrypt": "^5.0.1",
     "dotenv": "^10.0.0",
     "pg": "^8.7.1"
   }
